[facet][box] do not use modified elembctypes
This commit also modifies the tracer_tracer test, as this failed for box
after removing the elembctypes. This is because it was using Neumann
no-flow boundary conditions for the fracture domain on the left boundary
while the bulk problem used Dirichlet BCs. The old elembctypes for box
were discarding Dirichlet BCs on those nodes on the boundary that also
couple to fractures in order for the coupling conditions to be
fulfilled. However, that discards the boundary condition. With this
commit, Dirichlet BCs are always fulfilled, but for the tracer test this
means that as a consequence, the pressure in the fracture on the left
side is equal to the bulk pressure - because it uses Dirichlet-type
coupling conditions. This commit changes the test such that it uses
Dirichlet BCs in the bulk domain only on the lower part of the boundary
such that the fracture touches the domain boundary at a no-flow
boundary.